josbet-florist/requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
python-dotenv==1.0.0
mysql-connector-python==9.5.0
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
